Changed version to 0.0.4.
Enforce minimum version of 1.5.2 instead of the guide's advice of 1.4.0 for SimpleSchema, primarily because we use validator's "clean" option that was introduced in 1.5.0.

First implementation of testing. Just took ValidatedMethod's test cases, adjusted them to use this mixin, and then made sure they pass.

Explained in README that when validating with {}, the args to run and _execute must be defined. i.e. run({}) instead of run(). This may go away if ValidatedMethod makes an adjustment that they are contemplating.