Declaring the interface to be stable by switching to version 1.0.0 per SemVer.
Minor tweaks to the README to encourage reuse of Schema as opposed to duplicate specifications.

Note that one of the tests is failing. It will work as soon as ValidatedMethod starts adding a default doc when callers don't specify it. See https://github.com/meteor/validated-method/issues/39 for more info. Until then, just make sure you always call run with a empty options spec in the case of a ValidatedMethod that has no parameters. i.e. use methodname.run({}) instead of methodname.run().